feat: [CDS-56652]: move configFilesOutcome class
