+++ 125-cd-nextgen/modules/common/core/src/main/java/io/harness/cdng/configfile/steps/ConfigFilesStep.java
@@ -10,6 +10,7 @@ package io.harness.cdng.configfile.steps;
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.cdng.configfile.ConfigFileOutcome;
+import io.harness.cdng.configfile.ConfigFilesOutcome;
 import io.harness.cdng.service.steps.helpers.ServiceStepsHelper;
 import io.harness.cdng.stepsdependency.constants.OutcomeExpressionConstants;
 import io.harness.data.structure.EmptyPredicate;
