+++ 125-cd-nextgen/src/main/java/io/harness/serializer/kryo/NGKryoRegistrar.java
@@ -24,8 +24,8 @@ import io.harness.cdng.azure.webapp.ConnectionStringsParameters;
 import io.harness.cdng.azure.webapp.StartupCommandParameters;
 import io.harness.cdng.bamboo.BambooBuildStepInfo;
 import io.harness.cdng.chaos.ChaosStepNotifyData;
+import io.harness.cdng.configfile.ConfigFilesOutcome;
 import io.harness.cdng.configfile.steps.ConfigFileStepParameters;
-import io.harness.cdng.configfile.steps.ConfigFilesOutcome;
 import io.harness.cdng.customDeployment.FetchInstanceScriptStepInfo;
 import io.harness.cdng.customDeployment.FetchInstanceScriptStepParameters;
 import io.harness.cdng.elastigroup.ElastigroupBGStageSetupStepInfo;
