+++ 125-cd-nextgen/src/test/java/io/harness/cdng/configfile/ConfigFilesStepTest.java
@@ -20,7 +20,6 @@ import static org.mockito.Mockito.when;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.category.element.UnitTests;
 import io.harness.cdng.CDNGTestBase;
-import io.harness.cdng.configfile.steps.ConfigFilesOutcome;
 import io.harness.cdng.configfile.steps.ConfigFilesStep;
 import io.harness.cdng.service.steps.helpers.ServiceStepsHelper;
 import io.harness.pms.contracts.ambiance.Ambiance;
