+++ 125-cd-nextgen/src/test/java/io/harness/cdng/configfile/steps/ConfigFilesStepV2Test.java
@@ -26,6 +26,7 @@ import io.harness.cdng.CDStepHelper;
 import io.harness.cdng.configfile.ConfigFile;
 import io.harness.cdng.configfile.ConfigFileAttributes;
 import io.harness.cdng.configfile.ConfigFileWrapper;
+import io.harness.cdng.configfile.ConfigFilesOutcome;
 import io.harness.cdng.expressions.CDExpressionResolver;
 import io.harness.cdng.manifest.yaml.GitStore;
 import io.harness.cdng.manifest.yaml.storeConfig.StoreConfigType;
