+++ 125-cd-nextgen/src/test/resources/recaster/alias/alias.txt
@@ -372,7 +372,7 @@ io.harness.cdng.configfile.ConfigFile
 io.harness.cdng.configfile.ConfigFileOutcome
 io.harness.cdng.configfile.steps.ConfigFileStepParameters
 io.harness.cdng.configfile.ConfigFileAttributes
-io.harness.cdng.configfile.steps.ConfigFilesOutcome
+io.harness.cdng.configfile.ConfigFilesOutcome
 io.harness.cdng.manifest.yaml.OciHelmChartGenericConfig
 io.harness.cdng.manifest.yaml.harness.HarnessStore
 io.harness.cdng.gitops.steps.ClusterStepParameters
